+++ js/components/show-big-picture.js
@@ -1,41 +1,14 @@
+import { renderComments} from './render-comments.js';
+
 const body = document.body;
 const bigPicture = body.querySelector('.big-picture');
 const cancelButton = bigPicture.querySelector('.big-picture__cancel');
-
-
-const createComment = ({ avatar, name, message }, template) => {
-  const comment = template.cloneNode(true);
-
-  const image = comment.querySelector('.social__picture');
-  const socialText = comment.querySelector('.social__text');
-
-  image.src = avatar;
-  image.alt = name;
-  socialText.textContent = message;
-
-  return comment;
-};
-
-const renderComments = (comments) => {
-  const commentList = document.querySelector('.social__comments');
-  const template = commentList.children[0];
-
-  commentList.innerHTML = '';
-
-  const fragment = document.createDocumentFragment();
-  comments.forEach((comment) => {
-    const commentElement = createComment(comment, template);
-    fragment.append(commentElement);
-  });
-
-  commentList.append(fragment);
-};
-
-export { renderComments };
+const commentsLoader = bigPicture.querySelector('.comments-loader');
 
 const hideBigPicture = () => {
   bigPicture.classList.add('hidden');
   body.classList.remove('modal-open');
+  commentsLoader.classList.remove('hidden');
   document.removeEventListener('keydown', onEscKeyDown);
 };
 
@@ -50,15 +23,23 @@ const onCancelButtonClick = () => {
   hideBigPicture();
 };
 
-const showBigPicture = ({ url, likes, description, comments }) => {
+const showBigPicture = ({
+  url,
+  description,
+  comments,
+  likes
+}) => {
+
   const img = bigPicture.querySelector('.big-picture__img img');
-  const likesCount = bigPicture.querySelector('.likes-count');
   const socialCaption = bigPicture.querySelector('.social__caption');
+  const socialComment = bigPicture.querySelector('.social__comment-total-count');
+  const likesCount = bigPicture.querySelector('.likes-count');
 
   img.src = url;
-  img.alt = description;
+  socialCaption.alt = description;
+  socialComment.textContent = comments.length;
   likesCount.textContent = likes;
-  socialCaption.textContent = description;
+
 
   bigPicture.classList.remove('hidden');
   body.classList.add('modal-open');
@@ -70,4 +51,6 @@ const showBigPicture = ({ url, likes, description, comments }) => {
 
 cancelButton.addEventListener('click', onCancelButtonClick);
 
-export { showBigPicture };
+export {
+  showBigPicture, bigPicture
+};
